refactor(binding-kafka): move CreateTopics codec to exported api package
KafkaCreateTopicsRequestGenerator/KafkaCreateTopicsResponsePipeline lived
under internal.stream, which the Java module system doesn't export --
binding-mcp-kafka's upcoming create_topics tool needs to call them
directly. Move to a new io.aklivity.zilla.runtime.binding.kafka.api
package and export it; the public surface (generate()/decode() plus
records) was already clean, so this is a pure relocation.
