+++ runtime/binding-kafka/src/main/java/io/aklivity/zilla/runtime/binding/kafka/api/KafkaCreateTopicsResponsePipeline.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.aklivity.zilla.runtime.binding.kafka.internal.stream;
+package io.aklivity.zilla.runtime.binding.kafka.api;
 
 import java.nio.ByteOrder;
 import java.util.ArrayList;
