+++ runtime/binding-kafka/src/main/moditect/module-info.java
@@ -19,6 +19,8 @@ module io.aklivity.zilla.runtime.binding.kafka
     requires io.aklivity.zilla.runtime.engine;
     requires io.aklivity.zilla.config.binding.kafka;
 
+    exports io.aklivity.zilla.runtime.binding.kafka.api;
+    exports io.aklivity.zilla.runtime.binding.kafka.config;
     exports io.aklivity.zilla.runtime.binding.kafka.identity;
 
     uses io.aklivity.zilla.runtime.binding.kafka.identity.KafkaClientIdSupplierFactorySpi;
